replacing bootleg BigInt library with boost headers

+++ Euler_65.cpp
@@ -2,21 +2,21 @@
 
 #include "Euler.h"
 
-std::vector<BigInteger> recurseFraction2(std::vector<BigInteger> period, BigInteger n, std::vector<BigInteger> fraction)
+std::vector<cpp_int> recurseFraction2(std::vector<cpp_int> period, cpp_int n, std::vector<cpp_int> fraction)
 {
     if (n > period.size() - 1)
         return fraction;
 
     std::swap(fraction[0], fraction[1]);
 
-    fraction[0] = (fraction[1] * period[period.size() - n.toInt() - 1]) + fraction[0];
+    fraction[0] = (fraction[1] * period[period.size() - n - 1]) + fraction[0];
 
     return recurseFraction2(period, n + 1, fraction);
 }
 
-std::vector<BigInteger> recurseFraction2(std::vector<BigInteger> period, BigInteger n)
+std::vector<cpp_int> recurseFraction2(std::vector<cpp_int> period, cpp_int n)
 {
-    std::vector<BigInteger> fraction;
+    std::vector<cpp_int> fraction;
 
     fraction.push_back(period[period.size() - 1]);
     fraction.push_back(1);
@@ -24,9 +24,9 @@ std::vector<BigInteger> recurseFraction2(std::vector<BigInteger> period, BigInte
     return recurseFraction2(period, 1, fraction);
 }
 
-BigInteger periodiuy()
+cpp_int periodiuy()
 {
-    std::vector<BigInteger> period;
+    std::vector<cpp_int> period;
 
     period.push_back(2);
 
@@ -46,7 +46,7 @@ BigInteger periodiuy()
     }
 
     
-    std::vector<BigInteger> approx = recurseFraction2(period, 0);
+    std::vector<cpp_int> approx = recurseFraction2(period, 0);
     return approx[0];
 }
